--- .mocharc.json
@@ -1,10 +1,0 @@
-{
-  "exit": true,
-  "recursive": true,
-  "require": [
-    "ts-node/register",
-    "tsconfig-paths/register"
-  ],
-  "reporter": "mochawesome",
-  "timeout": "10000"
-}
