+++ package.json
@@ -15,7 +15,7 @@
         "lint:fix": "eslint --ext .ts --fix src/**",
         "lint:frontend": "eslint --ext .js src/main/assets/scripts/**",
         "lint:frontend:fix": "eslint --ext .js --fix src/main/assets/scripts/**",
-        "test": "mocha \"src/test/**/*.test.ts\" --reporter-options reportFilename=all,inlineAssets=true,reportTitle=CAS-buyer-frontend-tests-all",
+        "test": "env-cmd -f config/.env.test vitest run",
         "test:a11y": "mocha \"src/test/a11y/**/*.test.ts\" --timeout 60000 --reporter-options reportFilename=a11y,inlineAssets=true,reportTitle=CAS-buyer-frontend-a11y",
         "test:routes": "mocha \"src/test/routes/**/*.test.ts\" --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=CAS-buyer-frontend-routes",
         "test:unit": "env-cmd -f config/.env.test mocha \"src/test/unit/**/*.test.ts\" --reporter-options reportFilename=unit,inlineAssets=true,reportTitle=CAS-buyer-frontend-unit-tests-all",
@@ -45,7 +45,6 @@
         "@types/nunjucks": "^3.2.2",
         "@types/require-directory": "^2.1.2",
         "@types/serve-favicon": "^2.5.3",
-        "@types/sinon-chai": "^3.2.8",
         "axios": "^0.21.4",
         "ccs-frontend-kit": "^1.1.3",
         "config": "^3.3.9",
@@ -86,7 +85,6 @@
         "@babel/preset-env": "^7.22.5",
         "@types/chai-as-promised": "^7.1.5",
         "@types/jsdom": "^21.1.1",
-        "@types/mocha": "^9.0.0",
         "@types/pa11y": "^5.3.5",
         "@types/supertest": "^2.0.8",
         "@typescript-eslint/eslint-plugin": "^5.60.1",
@@ -104,21 +102,17 @@
         "gulp-livereload": "^4.0.2",
         "gulp-minify": "^3.1.0",
         "jsdom": "^22.1.0",
-        "mocha": "^6.2.2",
-        "mocha-junit-reporter": "^1.23.3",
-        "mocha-multi": "^1.1.3",
-        "mocha-parallel-tests": "^2.2.2",
-        "mochawesome": "^4.1.0",
+        "msw": "^1.2.2",
         "nock": "^13.3.1",
         "nyc": "^15.1.0",
         "pa11y": "^6.2.3",
         "prettier": "^2.8.8",
         "request": "^2.88.0",
-        "sinon": "^9.2.4",
-        "sinon-chai": "^3.7.0",
         "supertest": "^4.0.2",
         "ts-node-dev": "^2.0.0",
         "typescript": "^5.1.3",
-        "undici": "^5.22.1"
+        "undici": "^5.22.1",
+        "vite-tsconfig-paths": "^4.2.0",
+        "vitest": "^0.32.4"
     }
 }
