+++ src/specs/utils/momentFactory.spec.ts
@@ -1,0 +1,47 @@
+import { describe, it, expect, vi, beforeEach , afterEach } from 'vitest';
+import { MomentFactory } from '@utils/momentFactory';
+
+describe('Moment factory', () => {
+  describe('current date methods', () => {
+    const now = new Date('2023-05-03T14:45:00');
+
+    beforeEach(() => {
+      vi.useFakeTimers();
+      vi.setSystemTime(now);
+    });
+
+    afterEach(() => {
+      vi.useRealTimers();
+    });
+
+    it('returns 2023-05-03T14:45:00.000 for currentDateTime', () => {
+      expect(MomentFactory.currentDateTime().format('YYYY-MM-DDTHH:mm:ss.SSS')).to.eq('2023-05-03T14:45:00.000');
+    });
+
+    it('returns 2023-05-03T14:45:00.000 for currentDate', () => {
+      expect(MomentFactory.currentDate().format('YYYY-MM-DDTHH:mm:ss.SSS')).to.eq('2023-05-03T00:00:00.000');
+    });
+  });
+
+  describe('.maxDate', () => {
+    it('should return 9999-12-31T00:00:00.000', () => {
+      expect(MomentFactory.maxDate().format('YYYY-MM-DDTHH:mm:ss.SSS')).to.eq('9999-12-31T00:00:00.000');
+    });
+  });
+
+  describe('.parse', () => {
+    it('should throw an error when an empty string is parsed', () => {
+      expect(() => MomentFactory.parse('')).to.throw('Value must be defined');
+    });
+
+    it('returns 2023-05-04T00:00:00.000 when just the date is parsed', () => {
+      expect(MomentFactory.parse('2023-05-04').format('YYYY-MM-DDTHH:mm:ss.SSS')).to.eq('2023-05-04T00:00:00.000');
+    });
+
+    it('returns 2023-05-04T17:23:32.000 when just the date and time is parsed', () => {
+      expect(MomentFactory.parse('2023-05-04T17:23:32').format('YYYY-MM-DDTHH:mm:ss.SSS')).to.eq(
+        '2023-05-04T17:23:32.000'
+      );
+    });
+  });
+});
