+++ vite.config.ts
@@ -1,0 +1,17 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite';
+
+import viteTsConfigPaths from 'vite-tsconfig-paths';
+
+export default defineConfig({
+  plugins: [
+    viteTsConfigPaths()
+  ],
+
+  test: {
+    environment: 'node',
+    include: [
+      'src/specs/**/*.{test,spec}.ts'
+    ]
+  },
+});
